+++ src/hooks/admin/useAdminSanction.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import axios from 'axios';
 
-import { getCookies } from '@/utils/getCookies';
+import { getCookies } from '@/utils/cookie/getCookies';
 
 const useAdminSanction = (date: { memberId: string; reason: string }) => {
   const token = getCookies('jwt');
