+++ src/hooks/project/useProjectDetail.ts
@@ -1,8 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 
-import { getCookies } from '@/utils/getCookies';
-import { postCookies } from '@/utils/postCookies';
+import { getCookies } from '@/utils/cookie/getCookies';
+import { postCookies } from '@/utils/cookie/postCookies';
 
 const useProjectDetail = (id: string | undefined) => {
   const { isLoading, isError, data, isFetching } = useQuery({
